Changed coordinates from qint32 to quint32; fixed various .h issues; made Receiver and Sender abstract - deleting .cpp; Prepared .cpp files for implementation of what is now in .h files; Implemented the Matrix.

+++ trunk/Base.h
@@ -10,19 +10,19 @@
 
 class Base {
 public:
-    Base(quint8 color,qint32 x,qint32 y,quint32 width, quint32 height);
+    Base(quint8 color,quint32 x,quint32 y,quint32 width, quint32 height);
     virtual ~Base();
     
-    virtual float getCircleOverlap(qint32 x, qint32 y, quint32 radius) const; // vrati 0..1
+    virtual qreal getCircleOverlap(quint32 x, quint32 y, quint32 radius) const; // vrati 0..1
     virtual quint8 getColor() const {return color;}
 
 protected:
-    qint32 x1;
-    qint32 y1;
-    qint32 x2;
-    qint32 y2;
+    quint32 x1;
+    quint32 y1;
+    quint32 x2;
+    quint32 y2;
 
-    quint32 color;
+    quint8 color;
 
 private:
     Base(const Base& orig) {} // disabled
